==> notebooks/02) visualization.ipynb <==
### VISUALIZATION 

import pandas as pd
from datetime import timedelta 
import plotly.express as px
from typing import Optional


!pip install plotly --upgrade

tabular_data_final = pd.read_parquet('tabular_data_final.parquet')

tabular_data_final

### creating a plotting function 

def plot_one_sample(
    example_id: int,
    features: pd.DataFrame,
    targets: Optional[pd.Series] = None,
    predictions: Optional[pd.Series] = None
):
    """"""
    features_ = features.iloc[example_id]

    if targets is not None:
        target_ = targets.iloc[example_id]
    else:
        target_ = None

    ts_columns = [c for c in features.columns if c.startswith('rides_previous_')]
    ts_values = [features_[c] for c in ts_columns] + [target_]
    ts_dates = pd.date_range(
        features_['pickup_hour'] - timedelta(hours=len(ts_columns)),
        features_['pickup_hour'],
        freq='H'
    )

    # line plot with past values
    title = f'Pick up hour={features_["pickup_hour"]}, location_id={features_["pickup_location_id"]}'
    fig = px.line(
        x=ts_dates, y=ts_values,
        template='plotly_dark',
        markers=True, title=title
    )

    if targets is not None:
        # green dot for the value we wanna predict
        fig.add_scatter(x=ts_dates[-1:], y=[target_],
                        line_color='green',
                        mode='markers', marker_size=10, name='actual value')

    if predictions is not None:
        # big red X for the predicted value, if passed
        prediction_ = predictions.iloc[example_id]
        fig.add_scatter(x=ts_dates[-1:], y=[prediction_],
                        line_color='red',
                        mode='markers', marker_symbol='x', marker_size=15,
                        name='prediction')
    return fig

features = tabular_data_final.drop(columns=['target_rides_next_hour'])

targets = tabular_data_final['target_rides_next_hour']

plot_one_sample(
    example_id=0,
    features=features,
    targets=targets,
)

plot_one_sample(
    example_id=7,
    features=features,
    targets=targets,
)

plot_one_sample(
    example_id=760,
    features=features,
    targets=targets,
)
