==> notebooks/03) Baseline Model.ipynb <==
### MODELLING TABULAR DATA - + - TEST/TRAIN/SPLIT - + - XGBOODT/CATBOOST/LGM

from datetime import datetime 
from typing import Tuple 
import numpy as np 
import pandas as pd

df = pd.read_parquet('tabular_data_final.parquet')
df.head()

def train_test_split(
    df: pd.DataFrame,
    cutoff_date: datetime,
    target_column_name: str,
    ) -> Tuple[pd.DataFrame, pd.Series, pd.DataFrame, pd.Series]:
    """
    """
    train_data = df[df.pickup_hour < cutoff_date].reset_index(drop=True)
    test_data = df[df.pickup_hour >= cutoff_date].reset_index(drop=True)

    X_train = train_data.drop(columns=[target_column_name])
    y_train = train_data[target_column_name]
    X_test = test_data.drop(columns=[target_column_name])
    y_test = test_data[target_column_name]

    return X_train, y_train, X_test, y_test

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2023, 4, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}') 

#create a class (baselinemodelpreviousHour) - this function tells us the final values(last hour) of a locatiod ID on a particular date 

class BaselineModelPreviousHour:
    """
    Prediction = actual demand observed in the last hour
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test[f'rides_previous_1_hour']

model = BaselineModelPreviousHour()
prediction = model.predict(X_test)

prediction


from sklearn.metrics import mean_absolute_error

test_score = mean_absolute_error(y_test, prediction)
print(test_score)

from sklearn.linear_model import LinearRegression

# use only past rides data - create new indepeendent feature beacuse 'Pcikup_hour'(a feature) is in datetime format - and LR doesnt ingest date/time 
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

#model training 
model = LinearRegression()
model.fit(X_train_only_numeric, y_train)

#Removed the column in xtest aswell + 
X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions


x_test_only_numeric.head(2)

from sklearn.metrics import mean_absolute_error
test_score = mean_absolute_error(y_test, predictions)
print('test_score of baseline model:', test_score)

Our basleine Linear regaression model has a score of 5.92 - the models I plan of testing further on will be benchmarked against this
