==> notebooks/06) backfill feature store.ipynb <==
#!pip install hopsworks

from datetime import datetime
import pandas as pd
import hopsworks


import os
#from dotenv import load_dotenv
from src.paths import PARENT_DIR

# load key-value pairs from .env file located in the parent directory
#load_dotenv(PARENT_DIR / '.env')

HOPSWORKS_API_KEY = 'IB2b1Snz4GTx8CtU.XbeQ6PKTngkgDlSAmcgIMbZZKQErPI473MTYm9g2HC5maLTX9PyePJMWHcLWGbCh'

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
# to_year = datetime.now().year
to_year = 2023

print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

print(len(rides))

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

import hopsworks


connection = hopsworks.connection()


project = connection.create_project("Time_Series_NYC")


project = hopsworks.login(
    project = 'Time_Series_NYC',
    api_key_value= 'IB2b1Snz4GTx8CtU.XbeQ6PKTngkgDlSAmcgIMbZZKQErPI473MTYm9g2HC5maLTX9PyePJMWHcLWGbCh'

)

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

FEATURE_GROUP_NAME = 'new_taxi_time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = fs.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data, write_options={"wait_for_job": False})
